pyproject.toml
[project]
name = "k8s-schemas"
version = "0.1.0"
description = "Kubernetes CRD JSON Schema generator for IDE validation"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Gavin McFall"}
]
keywords = ["kubernetes", "crd", "json-schema", "yaml", "validation"]

dependencies = [
    "pyyaml>=6.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "jsonschema>=4.20.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://k8s-schemas.io"
Repository = "https://github.com/gavinmcfall/schema-gen"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
markers = [
    "integration: marks tests as integration tests (may require network)",
]
addopts = [
    "-v",
    "--tb=short",
]

[tool.coverage.run]
source = ["scripts"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by formatter)
]
